% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiedown_pdf_document.R
\name{uofgdataanalyticsreport}
\alias{uofgdataanalyticsreport}
\title{Customized R Markdown Document}
\usage{
uofgdataanalyticsreport(...)
}
\arguments{
\item{...}{Passed on to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
\code{uofgdataanalyticsreport()} is the main function of the uofgdataanalyticsreport package. Use as
\code{output: uofgdataanalyticsreport::uofgdataanalyticsreport} in the R Markdown YAML header.
}
